Cart functionality added with open/close using local state via apollo client

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -1,13 +1,14 @@
 // Where we interact with the Prisma database
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
-const { randomBytes } = require("crypto");
-const { promisify } = require("util");
-const { makeANiceEmail, transport } = require("../mail");
+const {randomBytes} = require("crypto");
+const {promisify} = require("util");
+const {makeANiceEmail, transport} = require("../mail");
+const {hasPermission} = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    if(!ctx.request.userId) {
+    if (!ctx.request.userId) {
       throw new Error("You must be logged in to do that");
     }
     const item = await ctx.db.mutation.createItem(
@@ -29,27 +30,31 @@ const Mutations = {
 
   updateItem(parent, args, ctx, info) {
     //copy of updates
-    const updates = { ...args };
+    const updates = {...args};
     delete updates.id;
     return ctx.db.mutation.updateItem(
       {
         data: updates,
-        where: { id: args.id }
+        where: {id: args.id}
       },
       info
     );
   },
 
   async deleteItem(parent, args, ctx, info) {
-    const where = { id: args.id };
+    const where = {id: args.id};
     //find the item
-    const item = await ctx.db.query.item({ where }, "{id, title}");
+    const item = await ctx.db.query.item({where}, "{id, title, user {id}}");
 
-    //check the user has the permissions to delete
-    //todo
+    //check the user has the permissions to delete - either they own the item or have delete permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
 
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permissions to delete");
+    }
     //delete it
-    return ctx.db.mutation.deleteItem({ where }, info);
+    return ctx.db.mutation.deleteItem({where}, info);
   },
   async signup(parent, args, ctx, info) {
     args.email = args.email.toLowerCase();
@@ -59,14 +64,14 @@ const Mutations = {
       {
         data: {
           ...args,
-          permissions: { set: ["USER"] }
+          permissions: {set: ["USER"]}
         }
       },
       info
     );
 
     //create jwt token for the user
-    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
+    const token = jwt.sign({userId: user.id}, process.env.APP_SECRET);
     //set jwt as cookie on the response
     console.log("In signup - the token is");
     console.log(token);
@@ -77,9 +82,9 @@ const Mutations = {
     return user;
   },
 
-  async signin(parent, { email, password }, ctx, info) {
+  async signin(parent, {email, password}, ctx, info) {
     //check if user exists with that email
-    const user = await ctx.db.query.user({ where: { email } });
+    const user = await ctx.db.query.user({where: {email}});
 
     if (!user) {
       throw new Error(`No such user found for email ${email}`);
@@ -90,7 +95,7 @@ const Mutations = {
       throw new Error("Invalid password");
     }
     //generate the jwt token
-    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
+    const token = jwt.sign({userId: user.id}, process.env.APP_SECRET);
     //set the cookie with the token
     console.log(token);
     ctx.response.cookie("token", token, {
@@ -102,12 +107,12 @@ const Mutations = {
   },
   signout(parent, args, ctx, info) {
     ctx.response.clearCookie("token");
-    return { message: "Goodbye" };
+    return {message: "Goodbye"};
   },
 
   async requestReset(parent, args, ctx, info) {
     //check if this is a real user
-    const user = await ctx.db.query.user({ where: { email: args.email } });
+    const user = await ctx.db.query.user({where: {email: args.email}});
 
     if (!user) {
       throw new Error(`No user with the email ${args.email}`);
@@ -118,7 +123,7 @@ const Mutations = {
     const resetTokenExpiry = Date.now() + 3600000; //1 hour from now
 
     const res = await ctx.db.mutation.updateUser({
-      where: { email: args.email },
+      where: {email: args.email},
       data: {
         resetToken: resetToken,
         resetTokenExpiry: resetTokenExpiry
@@ -134,10 +139,10 @@ const Mutations = {
         `Your password reset token is here! \n\n 
             <a href="${
           process.env.FRONTEND_URL
-        }/reset?resetToken=${resetToken}">Click here to reset</a> `
+          }/reset?resetToken=${resetToken}">Click here to reset</a> `
       )
     });
-    return { message: "Thanks!" };
+    return {message: "Thanks!"};
   },
   async resetPassword(parent, args, ctx, info) {
     //check if the passwords match
@@ -162,7 +167,7 @@ const Mutations = {
     //save new password to the user
     //remove old reset token fields
     const updatedUser = await ctx.db.mutation.updateUser({
-      where: { email: user.email },
+      where: {email: user.email},
       data: {
         password: password,
         resetToken: null,
@@ -171,7 +176,7 @@ const Mutations = {
     });
 
     //generate jwt
-    const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
+    const token = jwt.sign({userId: updatedUser.id}, process.env.APP_SECRET);
 
     //set the jwt cookie
     ctx.response.cookie("token", token, {
@@ -180,6 +185,39 @@ const Mutations = {
     });
     //return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    //check if user is logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
+
+    //query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+
+    //check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+
+    //update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {id: args.userId}
+      },
+      info
+    );
   }
 };
